tests(ui): add tests for consolidated connection validation
